requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
dnspython==2.3.0
filelock==3.10.0
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
grequests==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
mysql-connector-python==8.0.32
networkx==3.0
numpy==1.24.2
openai==0.27.2
pandas==2.0.0
pycparser==2.21
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
sympy==1.11.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
